refactor: reorganize statistics mod
Preaparing to introduce new changes in the repository.
