+++ src/servers/http/v1/services/scrape.rs
@@ -6,7 +6,7 @@
 //! and it returns the [`ScrapeData`] returned
 //! by the [`Tracker`].
 //!
-//! It also sends an [`statistics::Event`]
+//! It also sends an [`statistics::event::Event`]
 //! because events are specific for the HTTP tracker.
 use std::net::IpAddr;
 use std::sync::Arc;
@@ -48,10 +48,10 @@ pub async fn fake(tracker: &Arc<Tracker>, info_hashes: &Vec<InfoHash>, original_
 async fn send_scrape_event(original_peer_ip: &IpAddr, tracker: &Arc<Tracker>) {
     match original_peer_ip {
         IpAddr::V4(_) => {
-            tracker.send_stats_event(statistics::Event::Tcp4Scrape).await;
+            tracker.send_stats_event(statistics::event::Event::Tcp4Scrape).await;
         }
         IpAddr::V6(_) => {
-            tracker.send_stats_event(statistics::Event::Tcp6Scrape).await;
+            tracker.send_stats_event(statistics::event::Event::Tcp6Scrape).await;
         }
     }
 }
@@ -138,10 +138,10 @@ mod tests {
 
         #[tokio::test]
         async fn it_should_send_the_tcp_4_scrape_event_when_the_peer_uses_ipv4() {
-            let mut stats_event_sender_mock = statistics::MockEventSender::new();
+            let mut stats_event_sender_mock = statistics::event::sender::MockSender::new();
             stats_event_sender_mock
                 .expect_send_event()
-                .with(eq(statistics::Event::Tcp4Scrape))
+                .with(eq(statistics::event::Event::Tcp4Scrape))
                 .times(1)
                 .returning(|_| Box::pin(future::ready(Some(Ok(())))));
             let stats_event_sender = Box::new(stats_event_sender_mock);
@@ -150,7 +150,7 @@ mod tests {
                 Tracker::new(
                     &configuration::ephemeral().core,
                     Some(stats_event_sender),
-                    statistics::Repo::new(),
+                    statistics::repository::Repository::new(),
                 )
                 .unwrap(),
             );
@@ -162,10 +162,10 @@ mod tests {
 
         #[tokio::test]
         async fn it_should_send_the_tcp_6_scrape_event_when_the_peer_uses_ipv6() {
-            let mut stats_event_sender_mock = statistics::MockEventSender::new();
+            let mut stats_event_sender_mock = statistics::event::sender::MockSender::new();
             stats_event_sender_mock
                 .expect_send_event()
-                .with(eq(statistics::Event::Tcp6Scrape))
+                .with(eq(statistics::event::Event::Tcp6Scrape))
                 .times(1)
                 .returning(|_| Box::pin(future::ready(Some(Ok(())))));
             let stats_event_sender = Box::new(stats_event_sender_mock);
@@ -174,7 +174,7 @@ mod tests {
                 Tracker::new(
                     &configuration::ephemeral().core,
                     Some(stats_event_sender),
-                    statistics::Repo::new(),
+                    statistics::repository::Repository::new(),
                 )
                 .unwrap(),
             );
@@ -221,10 +221,10 @@ mod tests {
 
         #[tokio::test]
         async fn it_should_send_the_tcp_4_scrape_event_when_the_peer_uses_ipv4() {
-            let mut stats_event_sender_mock = statistics::MockEventSender::new();
+            let mut stats_event_sender_mock = statistics::event::sender::MockSender::new();
             stats_event_sender_mock
                 .expect_send_event()
-                .with(eq(statistics::Event::Tcp4Scrape))
+                .with(eq(statistics::event::Event::Tcp4Scrape))
                 .times(1)
                 .returning(|_| Box::pin(future::ready(Some(Ok(())))));
             let stats_event_sender = Box::new(stats_event_sender_mock);
@@ -233,7 +233,7 @@ mod tests {
                 Tracker::new(
                     &configuration::ephemeral().core,
                     Some(stats_event_sender),
-                    statistics::Repo::new(),
+                    statistics::repository::Repository::new(),
                 )
                 .unwrap(),
             );
@@ -245,10 +245,10 @@ mod tests {
 
         #[tokio::test]
         async fn it_should_send_the_tcp_6_scrape_event_when_the_peer_uses_ipv6() {
-            let mut stats_event_sender_mock = statistics::MockEventSender::new();
+            let mut stats_event_sender_mock = statistics::event::sender::MockSender::new();
             stats_event_sender_mock
                 .expect_send_event()
-                .with(eq(statistics::Event::Tcp6Scrape))
+                .with(eq(statistics::event::Event::Tcp6Scrape))
                 .times(1)
                 .returning(|_| Box::pin(future::ready(Some(Ok(())))));
             let stats_event_sender = Box::new(stats_event_sender_mock);
@@ -257,7 +257,7 @@ mod tests {
                 Tracker::new(
                     &configuration::ephemeral().core,
                     Some(stats_event_sender),
-                    statistics::Repo::new(),
+                    statistics::repository::Repository::new(),
                 )
                 .unwrap(),
             );
